refactor: muda script de popular dados de municípios para serviço
